requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-bootstrap5==23.4
django-css==2.3.1
django-htmx==1.18.0
django-phonenumber-field==7.1.0
django-qr-code==4.0.1
django-typed-models==0.14.0
html5lib==1.1
idna==3.7
lxml==5.2.2
oauthlib==3.2.2
oscrypto==1.3.0
phonenumbers==8.13.22
Pillow==10.0.1
platformdirs==4.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.5.1
python-environ==0.4.54
python3-openid==3.2.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
segno==1.6.1
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.14
